✈️ Migrate to TypeScript

+++ src/wizards/jokeWizard.ts
@@ -2,35 +2,38 @@ const { WizardScene } = require("telegraf/scenes");
 const stupidStuff_jokes = require("../data/stupidstuff.json");
 const reddit_jokes = require("../data/reddit_jokes.json");
 const wocka_jokes = require("../data/wocka.json");
+import { MyContext } from "../types";
+// import { CallbackQuery } from 'telegraf/types';
 
 const joke_providers = new Map([
   ["stupid", stupidStuff_jokes],
   ["wocka", wocka_jokes],
 ]);
 
-const STEP_1 = async (ctx) => {
-  inline_menu = [
+const STEP_1 = async (ctx:MyContext) => {
+
+  const inline_menu = [
     [{ text: "Reddit", callback_data: "reddit" }],
     [{ text: "Stupid Stuff", callback_data: "stupid" }],
     [{ text: "Wocka", callback_data: "wocka" }],
   ];
 
-  await ctx.reply("SELECT JOKE PROVIDER", {
-    reply_markup: { inline_keyboard: inline_menu },
-    reply_to_message_id: ctx.message.message_id,
-  });
+  await ctx.reply("SELECT JOKE PROVIDER", {reply_markup: { inline_keyboard: inline_menu }});
   return ctx.wizard.next();
 };
 
-const STEP_2 = async (ctx) => {
-  try {
-    if (ctx.updateType !== "callback_query") {
-      await ctx.reply("Invalid entry. Select only from the provided buttons");
-      return;
-    }
-    await ctx.answerCbQuery();
+const STEP_2 = async (ctx:MyContext) => {
+  if (ctx.updateType !== "callback_query") {
+    await ctx.reply("Invalid entry. Select only from the provided buttons");
+    return;
+  }
 
-    const user_selection = ctx.update.callback_query.data;
+  try {
+  await ctx.answerCbQuery();
+
+  if (("callback_query" in ctx.update) && ("data" in ctx.update.callback_query)) {
+
+    const user_selection = ctx.update.callback_query.data ;
 
     if (user_selection === "reddit") {
       await ctx.deleteMessage();
@@ -53,6 +56,12 @@ const STEP_2 = async (ctx) => {
       await ctx.reply("Invalid entry. Select only from the provided buttons");
       return;
     }
+  }
+  else {
+    await ctx.reply("Something went wrong try again 🤗.");
+    console.log("Message does not contain text");
+    return ctx.scene.leave();
+  }
   } catch (error) {
     await ctx.reply("Something went wrong try again. Try again 🤗");
     console.log(error);
@@ -63,4 +72,4 @@ const STEP_2 = async (ctx) => {
 
 const jokeWizard = new WizardScene("JOKE_WIZARD", STEP_1, STEP_2);
 
-module.exports = jokeWizard;
+export {jokeWizard};
